agents/requirements.txt
# AI Agent Dependencies
crewai>=0.1.0
firecrawl-py>=0.0.1
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
